Sort out dependencies, fix broken code and tests Remove botched ethcore-error crate
