+++ ethcore/src/lib.rs
@@ -88,6 +88,7 @@ extern crate parity_bytes as bytes;
 extern crate parity_crypto;
 extern crate parity_snappy as snappy;
 extern crate parking_lot;
+extern crate pod_account;
 extern crate trie_db as trie;
 extern crate patricia_trie_ethereum as ethtrie;
 extern crate rand;
@@ -98,6 +99,7 @@ extern crate parity_util_mem as malloc_size_of;
 extern crate rustc_hex;
 extern crate serde;
 extern crate stats;
+extern crate state_account;
 extern crate time_utils;
 extern crate triehash_ethereum as triehash;
 extern crate unexpected;
@@ -119,8 +121,6 @@ extern crate kvdb_rocksdb;
 extern crate blooms_db;
 #[cfg(any(test, feature = "env_logger"))]
 extern crate env_logger;
-#[cfg(test)]
-extern crate rlp_compress;
 #[cfg(test)]
 extern crate serde_json;
 
@@ -162,7 +162,6 @@ pub mod executive;
 pub mod machine;
 pub mod miner;
 pub mod pod_state;
-pub mod pod_account;
 pub mod snapshot;
 pub mod spec;
 pub mod state;
@@ -170,8 +169,8 @@ pub mod state_db;
 pub mod trace;
 pub mod transaction_ext;
 pub mod verification;
+pub mod account_db;
 
-mod account_db;
 mod externalities;
 mod factory;
 mod tx_filter;
